pyproject.toml
[project]
name = "adk-tutorial"
version = "0.1.0"
description = "Google ADK tutorial examples\""
authors = [
    {name = "pratik008",email = "pratik008@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk (>=0.1.0,<0.2.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
